requirements.txt
streamlit==1.31.0
google-generativeai==0.4.0
pandas==2.2.0
python-dotenv==1.0.1 